Move utilitarian methods into util.
